全面遷移 GUI Widget API 至 C++20 Modules
本次將原有以 .h 管理的 GUI Widget API 全面遷移至 C++20 Modules（.cppm），以 import/export 組織與引用各 Widget（Button、CheckBox、ListBox、Menu、Text 等），並移除大多數 .h 檔案。各 Widget 的 Private header 改為 import module，DIALOG.h 及 DIALOG_Intern.h 亦改用 import。main.cpp 及各 Widget .cpp 檔案皆以 import 取代 include。TUX.vcxproj 及 .filters 新增 .cppm 編譯規則並移除 .h。統一 Handle 型別為 WM_Obj*，bool 狀態 API 由 int/unsigned 改為 bool，item index/數量統一為 uint16_t。常數改為 constexpr 並 export，部分 API 參數型別更嚴謹，移除過時巨集，改用 constexpr 或 enum class。修正部分字元註解，統一以 module 管理依賴。此變更大幅現代化 API 結構，提升型別安全、可維護性與編譯效率。

+++ GUI/Widget/LISTBOX_Private.h
@@ -1,9 +1,10 @@
 #pragma once
 
-#include "LISTBOX.h"
-#include "WM.h"
-#include "GUI_ARRAY.h"
 #include "WIDGET.h"
+#include "GUI_ARRAY.h"
+#include "DIALOG_Intern.h"
+
+import TUX.Widget.ListBox;
 
 #define LISTBOX_ITEM_SELECTED (1 << 0)
 #define LISTBOX_ITEM_DISABLED (1 << 1)
@@ -42,10 +43,10 @@ struct LISTBOX_Obj : public WIDGET {
 	uint16_t ItemSpacing;
 };
 
-unsigned LISTBOX__GetNumItems(LISTBOX_Obj *pObj);
+uint16_t LISTBOX__GetNumItems(LISTBOX_Obj *pObj);
 const char *LISTBOX__GetpString(LISTBOX_Obj *pObj, int Index);
 void LISTBOX__InvalidateInsideArea(LISTBOX_Handle hObj);
 void LISTBOX__InvalidateItem(LISTBOX_Obj *pObj, int Sel);
 void LISTBOX__InvalidateItemAndBelow(LISTBOX_Obj *pObj, int Sel);
-void LISTBOX__InvalidateItemSize(LISTBOX_Obj *pObj, unsigned Index);
+void LISTBOX__InvalidateItemSize(LISTBOX_Obj *pObj, uint16_t Index);
 void LISTBOX__SetScrollbarWidth(LISTBOX_Obj *pObj);
